client-web: initial version
Add a minimal client to run in a web browser.

Note that unlike client-cli, it can not depend on artifex-rpc crate, as
this crate is built with tonic-build with "transport" feature enabled
(so we get ArtifexClient::connect()) but this fetches mio in such a way
that it can not be built for wasm32-unknown-unknown.

And as tonic-build is a build dependency, it will always be built for
the host, so toying with restricting dependencies/features according to
target_arch being/not being wasm32-unknown-unknown leads nowhere.

So client-web comes with its own build.rs and restricted features for
tonic and tonic-builder.