client-cli: add support for encrypted TLS client key
Add support for encrypted client private key, to connect via TLS, by:

- adding a password provider, able to read password from console (hidden input),
  a file, an environment variable or a file descriptor.
- adding a command line option ('-S') to artifex-client-cli to handle password.
- update tls::Config to support password.
- update ClientSigningKey and ClientCertResolver to be created using a builder
  pattern, thus allowing to optionally associate a password to the client
  private key file for decryption.

An empty password (e.g reading it from /dev/null) can be used if the client
private key file is not encrypted.

By default, the user is prompted for the password. To read password from file
"/some/path/password.txt", use:

```sh
artifex-client-cli -S file:/some/path/password.txt
```

Note: it is not possible to set the password in the "tls" section of the client
configuration file, as a security measure.