client-cli: tls: improve file-based TLS credentials handling
Instead of loading a certificate by using a plain url::Url, introduce
tls::uri::Uri and tls::file::FileUri to add support for password.

When using an item in an encrypted PEM file, it is now possible to pass the
password to decrypt it by using the "password" query parameter of the
URI (though passing password in clear text on the command line is a "Bad Idea"
®) or the source of the password using "password-source" query parameter (which
is more sensible).

The query parameter "password-source" accepts "env", "fd" and "file" sources.

A parsed URI with a password source can be converted to an URI with the password
sourced using Uri::source_secret().

Client key URI examples:

- file:///some/where/client.key.pem?password-source=env:CLIENT_KEY_PASSWORD
- file:///some/where/client.key.pem?password-source=fd:3
- file:///some/where/client.key.pem?password-source=file:/some/where/client.key.password.txt
- file:///some/where/client.key.pem?password=53cr3tP4ssw0rd